Moved a function to useful_functions.py to fix circular import (the function is not used anywhere yet)
