+++ petpal/utils/useful_functions.py
@@ -4,6 +4,7 @@ Module to handle abstracted functionalities
 import os
 import nibabel
 import numpy as np
+import pandas as pd
 from scipy.interpolate import interp1d
 import ants
 
@@ -304,3 +305,33 @@ def check_physical_space_for_ants_image_pair(image_1: ants.core.ANTsImage,
     org_cons = np.allclose(image_1.origin[:3], image_2.origin[:3])
 
     return dir_cons and spc_cons and org_cons
+
+
+def convert_ctab_to_dseg(ctab_path: str,
+                         dseg_path: str,
+                         column_names: list[str]=None):
+    """
+    Convert a FreeSurfer compatible color table into a BIDS compatible label
+    map ``dseg.tsv``.
+
+    Args:
+        ctab_path (str): Path to FreeSurfer compatible color table.
+        dseg_path (str): Path to ``dseg.tsv`` label mapfile to save.
+        column_names (list[str]): List of columns present in color table. Must
+            include 'mapping' and 'name'.
+    """
+    if column_names==None:
+        column_names = ['mapping','name','r','g','b','a','ttype']
+    fs_ctab = pd.read_csv(ctab_path,
+                          delim_whitespace=True,
+                          header=None,
+                          comment='#',
+                          names=column_names)
+    label_names = {'name': fs_ctab['name'],
+                   'mapping': fs_ctab['mapping'],
+                   'abbreviation': useful_functions.build_label_map(fs_ctab['name'])}
+    label_map = pd.DataFrame(data=label_names,
+                             columns=['name','abbreviation','mapping']).rename_axis('index')
+    label_map = label_map.sort_values(by=['mapping'])
+    label_map.to_csv(dseg_path,sep='\t')
+    return label_map
